Bot with difficulty level easy implemented with bugs
Simpler bot has been implemented with the bug that the player can sometimes select a card while the bot is running.
Other difficluty level are implemented but not tested and not running now.
Simple modifications on the game.service.ts for the bot implementation:
added variables :
pairsFoundPlayer
pairsFoundBot
botMemory
difficulty: 'easy' | 'medium' | 'hard' = 'easy'

change on Winning Message

A check Var (isPair) in the checkMacht function for the switch between Player and Bot